+++ MoodScape/Assets.xcassets/cAirForceBlue.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "161",
-          "green" : "137",
-          "red" : "102"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
