+++ MoodScape/Assets.xcassets/cBrownSugar.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "82",
-          "green" : "111",
-          "red" : "184"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
