+++ MoodScape/Assets.xcassets/cChineseViolet.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "129",
-          "green" : "93",
-          "red" : "153"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
