+++ MoodScape/Assets.xcassets/cCoral.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "100",
-          "green" : "135",
-          "red" : "247"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
