+++ MoodScape/Assets.xcassets/cJonquil.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "13",
-          "green" : "201",
-          "red" : "253"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
