+++ MoodScape/Assets.xcassets/cTiffanyBlue.colorset/Contents.json
@@ -11,24 +11,6 @@
         }
       },
       "idiom" : "universal"
-    },
-    {
-      "appearances" : [
-        {
-          "appearance" : "luminosity",
-          "value" : "dark"
-        }
-      ],
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "190",
-          "green" : "191",
-          "red" : "148"
-        }
-      },
-      "idiom" : "universal"
     }
   ],
   "info" : {
